Get list of drafts endpoint added with BDD tests
