+++ app/constants.py
@@ -19,3 +19,10 @@ MAX_STATUS_ACTOR_LEN = 100         # Maximum length of the actor column
 # Events Table Column Size Definitions
 
 MAX_EVENT_LEN = 20          # Maximum length of a event column
+
+# Endpoint names
+
+MESSAGE_BY_ID_ENDPOINT = "message"
+MESSAGE_LIST_ENDPOINT = "messages"
+
+MESSAGE_QUERY_LIMIT = 15
