+++ app/resources/messages.py
@@ -10,15 +10,13 @@ from app.common.labels import Labels
 from app.repository.modifier import Modifier
 from app.repository.retriever import Retriever
 from app.repository.saver import Saver
-from app.settings import MESSAGE_QUERY_LIMIT
 from app.validation.domain import MessageSchema
 from app.validation.user import User
 from app.resources.drafts import DraftModifyById
+from app.common.utilities import get_options, paginated_list_to_json
 
 logger = logging.getLogger(__name__)
 
-MESSAGE_LIST_ENDPOINT = "messages"
-MESSAGE_BY_ID_ENDPOINT = "message"
 
 """Rest endpoint for message resources. Messages are immutable, they can only be created."""
 
@@ -29,98 +27,17 @@ class MessageList(Resource):
     @staticmethod
     def get():
         """Get message list with options"""
-        string_query_args, page, limit, ru, survey, cc, label, business, desc = MessageList._get_options(request.args)
+        string_query_args, page, limit, ru, survey, cc, label, business, desc = get_options(request.args)
 
         message_service = Retriever()
         status, result = message_service.retrieve_message_list(page, limit, g.user_urn,
                                                                ru=ru, survey=survey, cc=cc, label=label,
                                                                business=business, descend=desc)
         if status:
-            resp = MessageList._paginated_list_to_json(result, page, limit, request.host_url,
+            resp = paginated_list_to_json(result, page, limit, request.host_url,
                                                        g.user_urn, string_query_args)
             resp.status_code = 200
             return resp
-
-    @staticmethod
-    def _get_options(args):
-        """extract options"""
-        string_query_args = '?'
-        page = 1
-        limit = MESSAGE_QUERY_LIMIT
-        ru = None
-        survey = None
-        cc = None
-        label = None
-        business = None
-        desc = True
-
-        if args.get('limit'):
-            limit = int(args.get('limit'))
-
-        if args.get('page'):
-            page = int(args.get('page'))
-
-        if args.get('ru'):
-            string_query_args = MessageList._add_string_query_args(string_query_args, 'ru', args.get('ru'))
-            ru = str(args.get('ru'))
-        if args.get('business'):
-            string_query_args = MessageList._add_string_query_args(string_query_args, 'business', args.get('business'))
-            business = str(args.get('business'))
-        if args.get('survey'):
-            survey = str(args.get('survey'))
-            string_query_args = MessageList._add_string_query_args(string_query_args, 'survey', args.get('survey'))
-        if args.get('cc'):
-            cc = str(args.get('cc'))
-            string_query_args = MessageList._add_string_query_args(string_query_args, 'cc', args.get('cc'))
-        if args.get('label'):
-            label = str(args.get('label'))
-            string_query_args = MessageList._add_string_query_args(string_query_args, 'label', args.get('label'))
-        if args.get('desc'):
-            desc = False if args.get('desc') == 'false' else True
-            string_query_args = MessageList._add_string_query_args(string_query_args, 'desc', args.get('desc'))
-
-        return string_query_args, page, limit, ru, survey, cc, label, business, desc
-
-    @staticmethod
-    def _add_string_query_args(string_query_args, arg, val):
-        """Create query string for get messages options"""
-        if string_query_args == '?':
-            return '{0}{1}={2}'.format(string_query_args, arg, val)
-        else:
-            return '{0}&{1}={2}'.format(string_query_args, arg, val)
-
-    @staticmethod
-    def _paginated_list_to_json(paginated_list, page, limit, host_url, user_urn, string_query_args):
-        """used to change a pagination object to json format with links"""
-        messages = []
-        msg_count = 0
-        arg_joiner = ''
-        if string_query_args != '?':
-            arg_joiner = '&'
-
-        for message in paginated_list.items:
-            msg_count += 1
-            msg = message.serialize(user_urn)
-            msg['_links'] = {"self": {"href": "{0}{1}/{2}".format(host_url, MESSAGE_BY_ID_ENDPOINT, msg['msg_id'])}}
-            messages.append(msg)
-
-        links = {
-            'first': {"href": "{0}{1}".format(host_url, "messages")},
-            'self': {"href": "{0}{1}{2}{3}page={4}&limit={5}".format(host_url, MESSAGE_LIST_ENDPOINT, arg_joiner,
-                                                                     string_query_args, page, limit)}
-        }
-
-        if paginated_list.has_next:
-            links['next'] = {
-                "href": "{0}{1}{2}{3}page={4}&limit={5}".format(host_url, MESSAGE_LIST_ENDPOINT, arg_joiner,
-                                                                string_query_args, (page + 1), limit)}
-
-        if paginated_list.has_prev:
-            links['prev'] = {
-                "href": "{0}{1}{2}{3}page={4}&limit={5}".format(host_url, MESSAGE_LIST_ENDPOINT, arg_joiner,
-                                                                string_query_args, (page - 1), limit)}
-
-        return jsonify({"messages": messages, "_links": links})
 
 
 class MessageSend(Resource):
