+++ app/settings.py
@@ -25,8 +25,6 @@ NOTIFICATION_TEMPLATE_ID = 'a1995c3d-68ce-42be-bddf-287b0870544b'
 NOTIFICATION_DEV_EMAIL = os.getenv('NOTIFICATION_DEV_EMAIL', 'gemma.irving@ons.gov.uk')
 
 
-MESSAGE_QUERY_LIMIT = os.getenv('MESSAGE_QUERY_LIMIT', 15)
-
 # SQLAlchemy configuration
 
 SQLALCHEMY_POOL_SIZE = os.getenv('SQLALCHEMY_POOL_SIZE', None)
